Refactor session storage into unified storage module
