+++ crates/goose/src/lib.rs
@@ -22,6 +22,7 @@ pub mod security;
 pub mod session;
 pub mod session_context;
 pub mod slash_commands;
+pub mod storage;
 pub mod subprocess;
 pub mod token_counter;
 pub mod tool_inspection;
